Update application.conf
Supposing we are on the last version of play framework and keeping in mind this: https://groups.google.com/g/play-framework/c/nQmUBoc97hQ

The fact that the cookies are signed protects you from session-hijacking: you can store the username in the session, and then trust that the user has not modified his session cookies to pretend to be he's someone else. What is not safe is information that has to be kept secret. For example, you should not store the credit card number in the session. 

Having the "secure" flag on the cookie would only mean that the browser will not send the cookie over http, only https. That would protect from eavesdropping on a local network, but would still leave an unencrypted trace on the user's computer so it's not recommended. 

.........

And the best case would be to update play framework version and edit conf file:

First of all, play framework version should be updated in order to fix the issue in conf file:

https://www.playframework.com/documentation/2.8.x/SettingsSession 
 secure = true
 maxAge = 2
 httpOnly = true